+++ modules/contracts/src.ts/services/ethService.ts
@@ -52,6 +52,7 @@ export const BIG_GAS_LIMIT = BigNumber.from(2_000_000);
 // nothing should ever be this expensive... _should_
 export const BIG_GAS_PRICE = parseUnits("1500", "gwei");
 
+// TODO: Deprecate. Note that this is used in autoRebalance.ts.
 export const waitForTransaction = async (
   provider: JsonRpcProvider,
   transactionHash: string,
@@ -114,7 +115,18 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     reason: TransactionReason,
     response: TransactionResponse,
   ): Promise<void> {
-    this.log.info({ method, methodId, channelAddress, reason, response }, "Tx submitted.");
+    this.log.info(
+      {
+        method,
+        methodId,
+        channelAddress,
+        reason,
+        hash: response.hash,
+        gasPrice: response.gasPrice.toString(),
+        nonce: response.nonce,
+      },
+      "Tx submitted.",
+    );
     await this.store.saveTransactionAttempt(onchainTransactionId, channelAddress, reason, response);
     this.evts[ChainServiceEvents.TRANSACTION_SUBMITTED].post({
       response: Object.fromEntries(
@@ -190,7 +202,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       async (): Promise<Result<TransactionResponse | undefined, Error>> => {
         try {
           // Send transaction using the passed in callback.
-          const actualNonce: number = nonce ?? (await signer.getTransactionCount());
+          const actualNonce: number = nonce ?? (await signer.getTransactionCount("pending"));
           const response: TransactionResponse | undefined = await txFn(gasPrice, actualNonce);
           return Result.ok(response);
         } catch (e) {
@@ -373,6 +385,10 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       try {
         /// SUBMIT
         // NOTE: Nonce will persist across iterations, as soon as it is defined in the first one.
+        this.log.debug(
+          { method, methodId, nonce, tryNumber, channelAddress, gasPrice: gasPrice.toString() },
+          "Attempting to send transaction",
+        );
         const result = await this.sendTx(txFn, gasPrice, signer, nonce);
         if (!result.isError) {
           const response = result.getValue();
@@ -386,6 +402,17 @@ export class EthereumChainService extends EthereumChainReader implements IVector
               // throw new ChainError(ChainError.reasons.)
             }
 
+            this.log.info(
+              {
+                hash: response.hash,
+                gas: response.gasPrice.toString(),
+                channelAddress,
+                method,
+                methodId,
+                nonce: response.nonce,
+              },
+              "Tx submitted",
+            );
             // Add this response to our local response history.
             responses.push(response);
             // NOTE: Response MUST be defined here because if it was NEVER defined (i.e. undefined on first iteration),
@@ -396,14 +423,21 @@ export class EthereumChainService extends EthereumChainReader implements IVector
           } else {
             // If response returns undefined, we assume the tx was not sent. This will happen if some logic was
             // passed into txFn to bail out at the time of sending.
-            this.log.warn({ method, methodId, channelAddress, reason }, "Did not attempt tx.");
+            this.log.info(
+              { method, methodId, channelAddress, reason, tryNumber, nonce, gasPrice: gasPrice.toString() },
+              "Did not attempt tx.",
+            );
             if (responses.length === 0) {
               // Iff this is the only iteration, then we want to go ahead return w/o saving anything.
+              this.log.info(
+                { method, methodId, channelAddress, tryNumber, nonce, gasPrice: gasPrice.toString() },
+                "Tx not needed",
+              );
               return Result.ok(undefined);
             } else {
               this.log.info(
-                { method, methodId, channelAddress, reason },
-                `txFn returned undefined on try ${tryNumber}. Proceeding to confirmation step.`,
+                { method, methodId, channelAddress, reason, tryNumber, nonce, gasPrice: gasPrice.toString() },
+                `txFn returned undefined, proceeding to confirmation step.`,
               );
             }
           }
@@ -423,12 +457,12 @@ export class EthereumChainService extends EthereumChainReader implements IVector
               error.message.includes("There is another transaction with same nonce in the queue."))
           ) {
             this.log.info(
-              { method, methodId, channelAddress, reason, nonce, error },
+              { method, methodId, channelAddress, reason, nonce, error: error.message },
               "Nonce already used: proceeding to check for confirmation in previous transactions.",
             );
           } else {
             this.log.error(
-              { method, methodId, channelAddress, reason, nonce, error },
+              { method, methodId, channelAddress, reason, nonce, tryNumber, gasPrice, error },
               "Error occurred while executing tx submit.",
             );
             throw error;
@@ -537,14 +571,20 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     // We must check for confirmation in all previous transactions. Although it's most likely
     // that it's the previous one, any of them could have been confirmed.
     const pollForReceipt = async (): Promise<TransactionReceipt | undefined> => {
+      // Save all reverted receipts for a check in case our Promise.race evaluates to be undefined.
+      let reverted: TransactionReceipt[] = [];
       // Make a pool of promises for resolving each receipt call (once it reaches target confirmations).
-      const response = await Promise.race<any>(
+      const receipt = await Promise.race<any>(
         responses
           .map((response) => {
             return new Promise(async (resolve) => {
               const r = await provider.getTransactionReceipt(response.hash);
-              if (r && r.confirmations >= numConfirmations) {
-                return resolve(r);
+              if (r) {
+                if (r.status === 0) {
+                  reverted.push(r);
+                } else if (r.confirmations >= numConfirmations) {
+                  return resolve(r);
+                }
               }
             });
           })
@@ -553,7 +593,15 @@ export class EthereumChainService extends EthereumChainReader implements IVector
           // and/or none of them have the number of confirmations we want.
           .concat(delay(2_000)),
       );
-      return response;
+      if (!!receipt) {
+        if (reverted.length === responses.length) {
+          // We know every tx was reverted.
+          // NOTE: The first reverted receipt in the array will be entirely arbitrary.
+          // TODO: Should we return the reverted receipt belonging to the latest tx instead?
+          return reverted[0];
+        }
+      }
+      return receipt;
     };
 
     // Poll for receipt.