+++ modules/test-runner/src/load/helpers/test.ts
@@ -1,10 +1,11 @@
-import { getRandomBytes32, RestServerNodeService } from "@connext/vector-utils";
+import { delay, getRandomBytes32, RestServerNodeService } from "@connext/vector-utils";
 import { constants } from "ethers";
 import PriorityQueue from "p-queue";
 
 import { env } from "../../utils";
 
-import { AgentManager } from "./agent";
+import { config } from "./config";
+import { AgentManager, TransferCompletedPayload } from "./agent";
 import { carolEvts, logger } from "./setupServer";
 
 export const cyclicalTransferTest = async (): Promise<void> => {
@@ -15,21 +16,47 @@ export const cyclicalTransferTest = async (): Promise<void> => {
     0,
   );
   const manager = await AgentManager.connect(agentService);
+  // console.log(manager)
 
   const killSwitch = await manager.startCyclicalTransfers();
 
   setTimeout(async () => {
     logger.warn({}, "Killing test");
     await killSwitch();
-  }, 90_000);
+    // wait 2s just in case
+    await delay(2_000);
+    process.exit(0);
+  }, config.testDuration);
+};
+
+/**
+ * Should create many transfers in a channel without ever
+ * resolving them.
+ */
+export const channelBandwidthTest = async (): Promise<void> => {
+  const agentService = await RestServerNodeService.connect(
+    env.carolUrl,
+    logger.child({ module: "RestServerNodeService" }),
+    carolEvts,
+    0,
+  );
+  const manager = await AgentManager.connect(agentService, false);
+
+  const killSwitch = await manager.createMultipleTransfersWithSameParties();
+
+  setTimeout(async () => {
+    logger.warn({}, "Killing test");
+    await killSwitch();
+    process.exit(0);
+  }, config.testDuration);
 };
 
 // Should create a bunch of transfers in the queue, with an
 // increasing concurrency
 export const concurrencyTest = async (): Promise<void> => {
   // Set test params
-  const maxConcurrency = 10;
-  const queuedPayments = 25; // added to queue
+  const maxConcurrency = config.maxConcurrency;
+  const queuedPayments = config.queuedPayments; // added to queue
 
   // Get agent manager
   let agentService: RestServerNodeService;
@@ -58,43 +85,69 @@ export const concurrencyTest = async (): Promise<void> => {
       .fill(0)
       .map((_) => {
         const [routingId, preImage] = [getRandomBytes32(), getRandomBytes32()];
-        manager.preImages[routingId] = preImage;
+        manager.transferInfo[routingId] = { ...(manager.transferInfo[routingId] ?? {}), preImage };
         return [routingId, preImage];
       });
 
   // Create tasks to fill queue with (25 random payments)
+  const minutesToWait = 20;
   const createTasks = () => {
     const paymentData = createPaymentData();
-    return Array(queuedPayments)
+    const tasks: [() => Promise<void>, Promise<TransferCompletedPayload | void>][] = Array(queuedPayments)
       .fill(0)
       .map((_, idx) => {
-        return async () => {
+        const [routingId, preImage] = paymentData[idx];
+        const creation = async () => {
           // Get random sender + receiver
           const sender = manager.getRandomAgent();
           const receiver = manager.getRandomAgent(sender);
 
-          // Save payment secrets to manager before creating
-          // payment
-          const [routingId, preImage] = paymentData[idx];
           await sender.createHashlockTransfer(receiver.publicIdentifier, constants.AddressZero, preImage, routingId);
           // NOTE: receiver will automatically resolve
         };
+        // wait 10min for completion
+        const completion = Promise.race([
+          manager.transfersCompleted.waitFor((data) => data.routingId === routingId),
+          delay(minutesToWait * 60 * 1000),
+        ]);
+        return [creation, completion];
       });
+    return tasks;
   };
-  let concurrency = 1;
+  let concurrency = 0;
+  let loopStats;
   for (const _ of Array(maxConcurrency).fill(0)) {
     // For loop runs one iteration of the test, with increasing
     // concurrency
+    concurrency += 1;
     logger.info({ concurrency }, "Beginning concurrency test");
     // Create a queue
     const queue = new PriorityQueue({ concurrency });
-    concurrency += 1;
 
-    const promises = createTasks().map((t) => queue.add(t));
+    const info = createTasks();
+    const creations = info.map(([t]) => queue.add(t));
 
-    await Promise.all(promises);
-    logger.info({}, "Test complete, increasing concurrency");
+    await Promise.all(creations);
+    logger.info({}, "Transfer creation complete, waiting for resolutions");
+
+    const completed = await Promise.all(info.map(([_, complete]) => complete));
+    const resolved = completed.filter((x) => !!x) as TransferCompletedPayload[];
+    const cancelled = resolved.filter((c) => c.cancelled);
+    loopStats = {
+      cancellationReasons: cancelled.map((c) => c.cancellationReason),
+      cancelled: cancelled.length,
+      resolved: resolved.length,
+      concurrency,
+    };
+    if (completed.length > resolved.length) {
+      logger.warn(
+        { ...loopStats, concurrency, completed: completed.length },
+        `Router can no longer forward within ${minutesToWait}min`,
+      );
+      break;
+    }
+    logger.info(loopStats, "Transfers resolved, increasing concurrency");
   }
-  logger.info({ concurrency: maxConcurrency, queuedPayments }, "Tests finished");
+  logger.info({ concurrency, maxConcurrency, queuedPayments, finalLoopStats: loopStats }, "Concurrency test finished");
   process.exit(0);
 };